Tugas Individu 6

+++ main/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.core import serializers
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse  
 from main.forms import ProductForm
 from main.models import Item
 from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
@@ -10,6 +10,9 @@ import datetime
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib import messages
+from django.views.decorators.csrf import csrf_exempt
+from django.views.decorators.http import require_POST
+import json
 
 @login_required(login_url='/login')
 def show_main(request):
@@ -32,17 +35,29 @@ def show_main(request):
 
     return render(request, "main.html", context)
 
+@csrf_exempt
+@require_POST
 def create_product(request):
-    form = ProductForm(request.POST or None)
+    name = request.POST.get("name");
+    description = request.POST.get("desc");
+    category = request.POST.get("category");
+    thumbnail = request.POST.get("thumbnail");
+    is_featured = request.POST.get("is_featured") == 'on';
+    price = request.POST.get("price");
+    user = request.user;
 
-    if form.is_valid() and request.method == "POST":
-        product_entry = form.save(commit = False)
-        product_entry.user = request.user
-        product_entry.save()
-        return redirect('main:show_main')
-    
-    context = {'form': form}
-    return render(request, "create_product.html", context)
+    new_product = Item(
+        name=name,
+        price=price,
+        description=description,
+        thumbnail=thumbnail,
+        is_featured=is_featured,
+        user=user
+    )
+
+    new_product.save();
+
+    return HttpResponse(b"CREATED", status=201);
 
 @login_required(login_url='/login')
 def show_product(request, product_id):
@@ -59,8 +74,21 @@ def show_xml(request):
 
 def show_json(request):
     product_list = Item.objects.all()
-    json_data = serializers.serialize("json", product_list)
-    return HttpResponse(json_data, content_type="application/json")
+    data = [
+        {
+            'id': str(item.id),
+            'name': item.name,
+            'price': item.price,
+            'description': item.description,
+            'thumbnail': item.thumbnail,
+            'category': item.category,
+            'is_featured': item.is_featured,
+            'terjual': item.terjual,
+            'user_id': item.user_id
+        }
+        for item in product_list
+    ]
+    return JsonResponse(data, safe=False)
 
 def show_xml_by_id(request, product_id):
     try:
@@ -72,11 +100,21 @@ def show_xml_by_id(request, product_id):
 
 def show_json_by_id(request, product_id):
     try:
-        product = Item.objects.get(pk = product_id)
-        json_data = serializers.serialize("json", [product])
-        return HttpResponse(json_data, content_type="application/json")
+        product = Item.objects.select_related('user').get(pk = product_id)
+        json_data = {
+            'id': str(product.id),
+            'name': product.name,
+            'price': product.price,
+            'description': product.description,
+            'thumbnail': product.thumbnail,
+            'category': product.category,
+            'is_featured': product.is_featured,
+            'terjual': product.terjual,
+            'user_id': product.user_id
+        }
+        return JsonResponse(json_data);
     except Item.DoesNotExist:
-        return HttpResponse(status=404)
+        return JsonResponse({'detail': 'Not Found'}, status=404);
 
 def register(request):
     form = UserCreationForm()
@@ -85,10 +123,16 @@ def register(request):
         form = UserCreationForm(request.POST)
         if form.is_valid():
             form.save()
-            messages.success(request, 'Your account has been successfully created!')
-            return redirect('main:login')
-    context = {'form': form}
-    return render(request, 'register.html', context)\
+            return JsonResponse({
+                'status': 'success',
+                'message': 'Account created! You will be redirected to login.'
+            }, status=201)
+        else:
+            errors = json.loads(form.errors.as_json())
+            return JsonResponse({'status': 'error', 'errors': errors}, status=400)
+    
+    return render(request, 'register.html')
+    
 
 def login_user(request):
     if request.method == "POST":
@@ -97,33 +141,86 @@ def login_user(request):
         if form.is_valid():
             user = form.get_user()
             login(request, user)
-            response = HttpResponseRedirect(reverse("main:show_main"))
+            response =  JsonResponse({
+                'status': 'success',
+                'message': 'Login succesfull!',
+                'redirect_url': reverse('main:show_main')
+            })
             response.set_cookie('last_login', str(datetime.datetime.now()))
             return response
-    else:
-        form = AuthenticationForm(request)
+        else:
+            errors = json.loads(form.errors.as_json())
+            return JsonResponse({'status': 'error', 'errors': errors}, status=400)
 
-    context= {'form': form}
-    return render(request, 'login.html', context)
+    return render(request, 'login.html')
 
+@csrf_exempt
 def logout_user(request):
     logout(request)
-    response = HttpResponseRedirect(reverse('main:login'))
-    response.delete_cookie('last_login')
-    return redirect('main:login')
 
+    redirect_url = reverse('main:login');
+
+    response = JsonResponse({
+        'status': 'success',
+        'message': 'You have been logged out.',
+        'redirect_url': redirect_url
+    })
+
+    response.delete_cookie('last_login');
+
+    return response
+
+@csrf_exempt
+@require_POST
 def edit_product(request, id):
     product = get_object_or_404(Item, pk = id)
-    form = ProductForm(request.POST or None, instance=product)
-    if form.is_valid() and request.method == "POST":
-        form.save()
-        return redirect('main:show_main')
-    
-    context = {'form': form}
 
-    return render(request, 'edit_product.html', context)
+    product.name = request.POST.get("name");
+    product.description = request.POST.get("desc");
+    product.category = request.POST.get("category");
+    product.thumbnail = request.POST.get("thumbnail");
+    product.is_featured = request.POST.get("is_featured") == 'on';
+    product.price = request.POST.get("price");
+    product.user = request.user;
 
+    product.save();
+
+    json_data = {
+        'id': str(product.id),
+        'name': product.name,
+        'price': product.price,
+        'description': product.description,
+        'thumbnail': product.thumbnail,
+        'category': product.category,
+        'is_featured': product.is_featured,
+        'terjual': product.terjual,
+        'user_id': product.user_id
+    }
+
+    return JsonResponse(json_data, status=201)
+
+@csrf_exempt
+@require_POST
 def delete_product(request, id):
     product = get_object_or_404(Item, pk = id)
+
+    if product.user != request.user:
+        return JsonResponse({'status': 'error', 'message': 'Not authorized'}, status=403)
+    
     product.delete()
-    return HttpResponseRedirect(reverse('main:show_main'))
+    return JsonResponse({'status': 'success', 'message': 'Product deleted successfully'})
+
+
+@login_required
+def get_page_data(request):
+    if request.user.is_authenticated:
+        return JsonResponse({
+            'is_authenticated': True,
+            'user_id': request.user.id,
+            'username': request.user.username,
+            'name': 'Prasetya Surya Syahputra',
+            'npm': '2406398381',
+            'class': 'PBP E',
+        })
+    else:
+        return JsonResponse({'is_authenticated': False})
